+++ src/clean_air/data/data_filter_handler.py
@@ -1,13 +1,9 @@
+#WIP . Need to add Chemical Species filters
+
 import os
 import yaml
 import glob
 import enum
-
-#Ansi code colours
-BLACK = "\033[0;30m"
-RED = "\033[0;31m"
-GREEN = "\033[0;32m"
-END = "\033[0m"
 
 
 def print_yaml_file(yaml_filename):
@@ -77,8 +73,6 @@ class DataFilterHandler:
 
         self.filters_dict[key] = True
 
-        print(f'{GREEN} Setting the filter to on for key: {key}, value: {self.filters_dict[key]}')
-        print(END)
 
     def turn_filter_off(self, key):
         """Turn the filter off for a given <yaml_filename> key
@@ -86,8 +80,6 @@ class DataFilterHandler:
 
         self.filters_dict[key] = False
 
-        print(f'{RED} Setting the filter to off for key: {key}, value: {self.filters_dict[key]}')
-        print(END)
 
     def is_filter_on(self, key)->bool:
         """Checks if filter is turned on for this dataset.
@@ -95,18 +87,11 @@ class DataFilterHandler:
 
         if key in self.filters_dict.keys():
             if self.filters_dict[key]:
-                #print("Filter On, ", "value =", self.filters_dict[key])
-                print(f"{GREEN} Filter ON, value = {self.filters_dict[key]}  \n")
                 return True
             else:
-                #print("Filter Off, ", "value =", self.filters_dict[key])
-                print(f"{RED} Filter Off, value = {self.filters_dict[key]}  \n")
                 return False
         else:
-            #print("Filter Off, No Key")
-            print(f"{RED} Filter Off, No Key  \n")
             return False
-        print(END)
 
     def get_filtered_data_subsets(self)->set:
         """ returns a set of keys <yaml_filename> for the filters that are switched on"""
@@ -116,10 +101,9 @@ class DataFilterHandler:
             if self.filters_dict[key]:
                 set_out.add(key)
 
-        print(set_out)
         return set_out
 
-    #TODO: if using the turn_filter_on_contains function, will need to rewrite to get to inner lists !!
+    #TODO: for CHEMICAL species ..if using the turn_filter_on_contains function, will need to rewrite to get to inner lists !!
 
     # def turn_filter_on_contains(self, filter_key, filter_value):
     #     """Turn the filter on for any keys that contain the <filter_key> , <filter_value>"""
@@ -149,80 +133,29 @@ class DataFilterHandler:
     def get_allfiles_dict(self):
         return self.yaml_allfiles_dict
 
-    # def print_access_inner_dict(self):
-    # #THINK THIS CRAZY BIT OF CODE ACTUALLY DOES WHAT I WANT :-)
-    # # HAVE TWO VERSIONS SO CAN UPDATE THE FILTERS ( PLAIN & INNER VERSIONS)
-    # #KEEP THIS PRINT VERSION TOO SO CAN SEE WHAT HECK IS GOING ON !!
-    #     for outer_key in self.get_allfiles_dict():
-    #         print('Outer Key = ', outer_key)  #outer key is the yaml filename
-    #         for inner_key in self.get_allfiles_dict()[outer_key]: #Now we have a possible key/value pair (this enough for simple ones)
-    #             print('   ', 'Inner Key', inner_key, 'Inner Value', self.get_allfiles_dict()[outer_key][inner_key])
-    #             #Now check to see if inner key is actually a list of key/value pairs
-    #             if type(self.get_allfiles_dict()[outer_key][inner_key]) is list:
-    #                 for goal in self.get_allfiles_dict()[outer_key][inner_key]:
-    #                     print('      ', 'List' , goal, type(goal))
-    #                     #Now get the values in these mini dictionaries
-    #                     if type(goal) is dict:
-    #                         for key, value in goal.items():
-    #                             print(str(key) + " => " + str(value))
 
+
+    # TODO: for CHEMICAL SPECIES ..this inner dictionary access not currently being used but is to help me with chemical species filter
     def print_access_inner_dict(self):
     #THINK THIS CRAZY BIT OF CODE ACTUALLY DOES WHAT I WANT :-)
     # PUT THIS IN THE FILTER HANDLER , NOT HERE ..AND HAVE TWO VERSIONS SO CAN UPDATE THE FILTERS ( PLAIN & INNER VERSIONS)
     #KEEP THIS PRINT VERSION TOO SO CAN SEE WHAT HECK IS GOING ON !!
         for outer_key in self.get_allfiles_dict():
-            print('Outer Key = ', outer_key)  #outer key is the yaml filename
+            # print('Outer Key = ', outer_key)  #outer key is the yaml filename
             for inner_key in self.get_allfiles_dict()[outer_key]: #Now we have a possible key/value pair (this enough for simple ones)
-                print('   ',
-                      'Inner Key', inner_key,
-                      'Inner Value', self.get_allfiles_dict()[outer_key][inner_key] ,
-                      'Inner Value Type', type(self.get_allfiles_dict()[outer_key][inner_key]))
+                #print('   ',
+                #      'Inner Key', inner_key,
+                #      'Inner Value', self.get_allfiles_dict()[outer_key][inner_key] ,
+                #      'Inner Value Type', type(self.get_allfiles_dict()[outer_key][inner_key]))
+
                 #Now check to see if inner key is actually a list of key/value pairs
                 if type(self.get_allfiles_dict()[outer_key][inner_key]) is list:
                    for goal in self.get_allfiles_dict()[outer_key][inner_key]:
-                       print('      ', type(self.get_allfiles_dict()[outer_key][inner_key]) , goal, type(goal))
+                       #print('      ', type(self.get_allfiles_dict()[outer_key][inner_key]) , goal, type(goal))
                        #Now get the values in these mini dictionaries
                        if type(goal) is dict:
                           for key, value in goal.items():
-                              print('         ', type(goal), str(key) + " => " + str(value))
+                            #print('         ', type(goal), str(key) + " => " + str(value))
+                            pass
 
-
-    def test(self):
-
-        # print yaml all file dict
-        print("print yaml all file dict")
-        dict_printer(self.yaml_allfiles_dict)
-        dict_printer(self.filters_dict)
-
-        print("now check if a filter is on")
-        # now check if a filter is on:
-        self.is_filter_on('/net/home/h05/clucas/PycharmProjects/CleanAirProject/clean_air/data/music.yaml')
-        self.is_filter_on('/net/home/h05/clucas/PycharmProjects/CleanAirProject/clean_air/data/dance.yaml')
-
-        # Turn a filter off
-        print("Turn a filter off")
-        self.turn_filter_off('/net/home/h05/clucas/PycharmProjects/CleanAirProject/clean_air/data/fruits.yaml')
-        self.turn_filter_off('/net/home/h05/clucas/PycharmProjects/CleanAirProject/clean_air/data/veggies.yaml')
-
-
-        self.is_filter_on('/net/home/h05/clucas/PycharmProjects/CleanAirProject/clean_air/data/fruits.yaml')
-        dict_printer(self.filters_dict)
-
-        print("Turn a filter on")
-        self.turn_filter_on('/net/home/h05/clucas/PycharmProjects/CleanAirProject/clean_air/data/fruits.yaml')
-        self.is_filter_on('/net/home/h05/clucas/PycharmProjects/CleanAirProject/clean_air/data/fruits.yaml')
-        dict_printer(self.filters_dict)
-
-        #self.turn_filter_off_contains('filetype','ground')
-        #self.turn_filter_on_contains('sweet_potato','2')
-        #self.turn_filter_on_contains('chemical_species','no2')
-
-        print("GET FILTERED DATA SUBSETS")
-        self.get_filtered_data_subsets()
-
-    # Run main
-
-
-#d = DataFilterHandler()
-#d.test()
