--- src/clean_air/data/fruits.yaml
@@ -1,5 +1,0 @@
-apples: 20
-mangoes: 2
-bananas: 3
-grapes: 100
-pineapples: 1
