--- src/clean_air/data/veggies.yaml
@@ -1,5 +1,0 @@
-brocolli: 20
-sweet_potato: 2
-carrots: 3
-swede: 100
-savoy_cabbage: 1
